+++ src/effects/vision5e.js
@@ -1,5 +1,5 @@
 import { effectModules } from "./effects.js";
-import { baseSpellEffect } from "./specialSpells";
+import { baseSpellEffect } from "./specialSpells.js";
 
 
 const VISION_EFFECTS = {
@@ -28,6 +28,11 @@ const VISION_EFFECTS = {
     type: "spell",
     transfer: false,
   },
+  "Eldritch Invocations: Devil's Sight": {
+    effectName: "Devil's Sight",
+    type: "feat",
+    transfer: true,
+  },
   "Devil's Sight": {
     effectName: "Devil's Sight",
     type: "feat",
@@ -51,7 +56,10 @@ export function addVision5eStub(document) {
   const name = document.flags.ddbimporter?.originalName ?? document.name;
 
   // if document name in Vision effects then add effect
-  if (VISION_EFFECTS[name] && document.type === VISION_EFFECTS[name].type) {
+  if (VISION_EFFECTS[name]
+    && document.type === VISION_EFFECTS[name].type
+    && !document.effects.some((e) => (e.name ?? e.label) === VISION_EFFECTS[name].effectName)
+  ) {
     const effect = baseSpellEffect(document, VISION_EFFECTS[name].effectName);
     effect.transfer = VISION_EFFECTS[name].transfer;
     document.effects.push(effect);
@@ -59,7 +67,7 @@ export function addVision5eStub(document) {
   return document;
 }
 
-export async function addVision5eStubs(documents) {
+export function addVision5eStubs(documents) {
   // check that we can gen effects
   const deps = effectModules();
   if (!deps.vision5eInstalled) return documents;