+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 - The check to copy the flag over to the new item to see if it should ignore a Chris's Premades was happening after Chris's Premades have been applied.
 - If Vision/Detection Modes 5e module is installed, DDB Importer will add stub effects to feats and spells it knows about for the module to use.
+- Status effect tweaks in v11.
 
 # 3.4.34